refactor(api): move subscription in-memory deps into bootstrap module
